requirements-cloud.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
pdfminer.six==20260107
faiss-cpu==1.13.2
numpy==2.4.1
prometheus_client==0.24.1
openai==2.15.0
google-genai==1.62.0
requests==2.32.5
